requirements.txt
opencv-python
torch
torchvision
joblib
pygame
streamlit
ultralytics
Pillow
numpy
av
streamlit-webrtc
